Update README.md
This School Billing System for students fees will help the user to Pay school bills online and print term reports. There are different types of users like teacher, student and admin. The student shall be able to print reports for the fee paid. He shall be able to print reports of the term specified. The teacher shall be able to maintain the data of each student. The system will also notify the students of outstanding fee payments.